requirements.txt
networkx==2.4
numpy==1.17.4
scikit-learn==0.21.3
scipy==1.3.3
seaborn==0.9.0
tensorflow-gpu==2.0.0a0
